+++ feature/fortune/src/main/java/org/sopt/official/feature/fortune/FoundationScreen.kt
@@ -37,8 +37,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
 import org.sopt.official.designsystem.SoptTheme
 import org.sopt.official.feature.fortune.component.FortuneTopBar
-import org.sopt.official.feature.fortune.feature.fortundDetail.navigation.FortuneDetail
-import org.sopt.official.feature.fortune.feature.fortundDetail.navigation.fortuneDetailNavGraph
+import org.sopt.official.feature.fortune.feature.fortuneDetail.navigation.FortuneDetail
+import org.sopt.official.feature.fortune.feature.fortuneDetail.navigation.fortuneDetailNavGraph
 import org.sopt.official.feature.fortune.feature.fortuneAmulet.navigation.FortuneAmulet
 import org.sopt.official.feature.fortune.feature.fortuneAmulet.navigation.fortuneAmuletNavGraph
 import org.sopt.official.feature.fortune.feature.home.navigation.Home
